+++ src/init.c
@@ -54,8 +54,8 @@ extern SEXP vctrs_chop_seq(SEXP, SEXP, SEXP, SEXP);
 extern r_obj* ffi_slice_seq(r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_slice_rep(r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_vec_restore(r_obj*, r_obj*);
+extern r_obj* ffi_vec_restore_recurse(r_obj*, r_obj*);
 extern r_obj* ffi_vec_restore_default(r_obj*, r_obj*);
-extern SEXP vec_proxy(SEXP);
 extern SEXP vec_proxy_equal(SEXP);
 extern SEXP vec_proxy_compare(SEXP);
 extern SEXP vec_proxy_order(SEXP);
@@ -223,8 +223,10 @@ static const R_CallMethodDef CallEntries[] = {
   {"ffi_slice_seq",                         (DL_FUNC) &ffi_slice_seq, 4},
   {"ffi_slice_rep",                         (DL_FUNC) &ffi_slice_rep, 3},
   {"ffi_vec_restore",                       (DL_FUNC) &ffi_vec_restore, 2},
+  {"ffi_vec_restore_recurse",               (DL_FUNC) &ffi_vec_restore_recurse, 2},
   {"ffi_vec_restore_default",               (DL_FUNC) &ffi_vec_restore_default, 2},
-  {"vctrs_proxy",                           (DL_FUNC) &vec_proxy, 1},
+  {"ffi_vec_proxy",                         (DL_FUNC) &vec_proxy, 1},
+  {"ffi_vec_proxy_recurse",                 (DL_FUNC) &vec_proxy_recurse, 1},
   {"vctrs_proxy_equal",                     (DL_FUNC) &vec_proxy_equal, 1},
   {"vctrs_proxy_compare",                   (DL_FUNC) &vec_proxy_compare, 1},
   {"vctrs_proxy_order",                     (DL_FUNC) &vec_proxy_order, 1},