+++ src/proxy-restore.h
@@ -4,11 +4,20 @@
 #include "vctrs-core.h"
 
 
-r_obj* vec_restore(r_obj* x, r_obj* to, const enum vctrs_owned owned);
-r_obj* vec_restore_default(r_obj* x, r_obj* to, const enum vctrs_owned owned);
+r_obj* vec_restore(r_obj* x, r_obj* to, enum vctrs_owned owned);
+r_obj* vec_restore_default(r_obj* x, r_obj* to, enum vctrs_owned owned);
 
-r_obj* vec_bare_df_restore(r_obj* x, r_obj* to, const enum vctrs_owned owned);
-r_obj* vec_df_restore(r_obj* x, r_obj* to, const enum vctrs_owned owned);
+r_obj* vec_restore_recurse(r_obj* x, r_obj* to, enum vctrs_owned owned);
+
+r_obj* vec_df_restore(r_obj* x,
+                      r_obj* to,
+                      enum vctrs_owned owned,
+                      enum vctrs_recurse recurse);
+
+r_obj* vec_bare_df_restore(r_obj* x,
+                           r_obj* to,
+                           enum vctrs_owned owned,
+                           enum vctrs_recurse recurse);
 
 
 #endif