+++ src/slice-assign.c
@@ -143,6 +143,12 @@ r_obj* vec_proxy_assign_opts(r_obj* proxy,
                              const struct vec_assign_opts* opts) {
   int n_protect = 0;
 
+  // Ignore vectors marked as fallback because the caller will apply
+  // a fallback method instead
+  if (vec_is_common_class_fallback(proxy)) {
+    return proxy;
+  }
+
   struct vec_assign_opts mut_opts = *opts;
   bool ignore_outer_names = mut_opts.ignore_outer_names;
   mut_opts.ignore_outer_names = false;
@@ -256,7 +262,9 @@ r_obj* raw_assign(r_obj* x, r_obj* index, r_obj* value, const enum vctrs_owned o
   int* index_data = r_int_begin(index);                                 \
                                                                         \
   if (n != r_length(value)) {                                           \
-    r_stop_internal("`value` should have been recycled to fit `x`.");   \
+    r_stop_internal("`value` (size %d) doesn't match `x` (size %d).",   \
+                    r_length(value),                                    \
+                    n);                                                 \
   }                                                                     \
                                                                         \
   r_obj* out = KEEP(vec_clone_referenced(x, owned));                    \
@@ -278,7 +286,9 @@ r_obj* raw_assign(r_obj* x, r_obj* index, r_obj* value, const enum vctrs_owned o
   r_ssize step = index_data[2];                                         \
                                                                         \
   if (n != r_length(value)) {                                           \
-    r_stop_internal("`value` should have been recycled to fit `x`.");   \
+    r_stop_internal("`value` (size %d) doesn't match `x` (size %d).",   \
+                    r_length(value),                                    \
+                    n);                                                 \
   }                                                                     \
                                                                         \
   r_obj* out = KEEP(vec_clone_referenced(x, owned));                    \
@@ -347,12 +357,17 @@ r_obj* df_assign(r_obj* x,
 
     // No need to cast or recycle because those operations are
     // recursive and have already been performed. However, proxy and
-    // restore are not recursive so need to be done for each element
-    // we recurse into. `vec_proxy_assign()` will proxy the `value_elt`.
-    r_obj* proxy_elt = KEEP(vec_proxy(out_elt));
+    // restore are not necessarily recursive and we might need to
+    // proxy each element we recurse into.
+    //
+    // NOTE: `vec_proxy_assign()` proxies `value_elt`.
+    r_obj* proxy_elt = KEEP(opts->recursive ? out_elt : vec_proxy(out_elt));
 
     r_obj* assigned = KEEP(vec_proxy_assign_opts(proxy_elt, index, value_elt, owned, opts));
-    assigned = vec_restore(assigned, out_elt, owned);
+
+    if (!opts->recursive) {
+      assigned = vec_restore(assigned, out_elt, owned);
+    }
 
     r_list_poke(out, i, assigned);
     FREE(2);